feat: settings scaffold
